+++ apiphany-core/src/main/java/org/apiphany/lang/LibraryDescriptor.java
@@ -2,6 +2,8 @@ package org.apiphany.lang;
 
 import java.util.Objects;
 
+import org.morphix.reflection.Reflection;
+
 /**
  * Descriptor for a library indicating its presence and associated specific class.
  *
@@ -14,7 +16,7 @@ public class LibraryDescriptor<T> {
 	/**
 	 * Indicates if the library is present and the specific class associated with it.
 	 */
-	private final boolean present;
+	private final boolean libraryPresent;
 
 	/**
 	 * The specific class associated with the library.
@@ -27,9 +29,33 @@ public class LibraryDescriptor<T> {
 	 * @param present indicates if the library is present
 	 * @param specificClass the specific class associated with the library
 	 */
-	private LibraryDescriptor(final boolean present, final Class<T> specificClass) {
-		this.present = present;
+	protected LibraryDescriptor(final boolean present, final Class<T> specificClass) {
+		this.libraryPresent = present;
 		this.specificClass = Objects.requireNonNull(specificClass, "specificClass must not be null");
+	}
+
+	/**
+	 * Factory method to create a {@link LibraryDescriptor} instance indicating the library is present.
+	 *
+	 * @param <T> the type of the specific class
+	 *
+	 * @param specificClass the specific class associated with the library
+	 * @return a new LibraryDescriptor instance with presence set to true
+	 */
+	public static <T> LibraryDescriptor<T> present(final Class<T> specificClass) {
+		return new LibraryDescriptor<>(true, specificClass);
+	}
+
+	/**
+	 * Factory method to create a {@link LibraryDescriptor} instance indicating the library is not present.
+	 *
+	 * @param <T> the type of the specific class
+	 *
+	 * @param specificClass the specific class associated with the library
+	 * @return a new LibraryDescriptor instance with presence set to false
+	 */
+	public static <T> LibraryDescriptor<T> notPresent(final Class<T> specificClass) {
+		return new LibraryDescriptor<>(false, specificClass);
 	}
 
 	/**
@@ -37,12 +63,14 @@ public class LibraryDescriptor<T> {
 	 *
 	 * @param <T> the type of the specific class
 	 *
-	 * @param present indicates if the library is present
+	 * @param libraryClassName the fully qualified class name to check for presence
 	 * @param specificClass the specific class associated with the library
 	 * @return a new LibraryDescriptor instance
 	 */
-	public static <T> LibraryDescriptor<T> of(final boolean present, final Class<T> specificClass) {
-		return new LibraryDescriptor<>(present, specificClass);
+	public static <T> LibraryDescriptor<T> of(final String libraryClassName, final Class<T> specificClass) {
+		return Reflection.isClassPresent(libraryClassName)
+				? present(specificClass)
+				: notPresent(specificClass);
 	}
 
 	/**
@@ -50,8 +78,8 @@ public class LibraryDescriptor<T> {
 	 *
 	 * @return true if the library is present, false otherwise
 	 */
-	public boolean isPresent() {
-		return present;
+	public boolean isLibraryPresent() {
+		return libraryPresent;
 	}
 
 	/**